Move MoveArrows.cs to its own directory
